refactor: simplify CustomAgentforceView to minimal JSON.stringify example
Move the complex recursive key/value table rendering to
ComplexAgentforceDataView.tsx. CustomAgentforceView is now a minimal
example (~90 lines) that shows the ViewProvider wiring clearly:
badge + definition + JSON.stringify(properties).
